Add registration push to share
